Start Object transfering
