+++ .gitignore
@@ -22,6 +22,7 @@
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 *.xml
+*.iml
 .idea/
 production/
-*.iml
+Project Complete/
